Refactor neighborhood methods
